Enable context in tracebacks for code defined interactively
We have to define a distinct filename for each compiled code ("<mypython-N>"
where N is the prompt number, instead of just "<mypython>"). In doctest mode,
the filename is still "<stdin>", meaning this feature doesn't work there
(which matches the default Python interpreter anyway).

In the tests, add the functionality to set the prompt number in _test_output,
which must be done for multiple statements for the tracebacks to work
properly. This needs to be abstracted better in the future.

Also fix a test that should have had a newline.