Use the PuDB _wait_for_mainpyfile functionality when it works
Because we compile the code in parts when it ends in an expression, this
causes pudb to exit without executing the expression part for a multiline
statement that ends in an expression. So in that case, we have to keep using
what we were using.

Also, in each case, PuDB annoyingly has a "Return: None" before it exits (with
the expression at the end of a multiline statement, it does it before the
expression as well).