Use separate globals and locals dictionaries
This allows us to store things like In, Out, _, __, and ___, and possibly
other things in the future, in globals(), and not have to worry about
reassigning them in the session, since all variable assignments in the session
will go to locals.

It's worth noting that to work properly functions that only allow one of
globals/locals to be passed in should pass in locals (even if it calls it
globals), such as timeit.Timer. That way they actually get the variables
defined interactively. globals() will generally just be the builtin names +
the mypython builtins like In, etc.

This fixes reassigning 'In' breaking mypython (you can still break it if you
change globals()['In'], but that's effectively monkeypatching at that point).