Use SetPropagatingDict in pudb_shell.py
Otherwise things like

a = 1
def test():
    print(a)

test()

don't act as expected, because the assignment to a is only set in locals().
This is because in a debugging context, at least when not at the module level,
locals() and globals() are different, which is different from the usual way
the interactive shell works.