Clean up traceback line hiding logic
Now we subclass the FormatException class from the standard library. This lets
us detect mypython frames exactly, instead of just by string matching,
both for purposes of removing them from the traceback and for warning about
errors from mypython.

Also only hide the exceptions if the first part of the exception is mypython
dirs. If mypython dirs exist after the first part, or if no non-mypython dirs
exist in the traceback, always show it as an error from mypython.

Hopefully this logic doesn't need to be modified in case of callbacks from the
standard library (I can't think of any at the moment).