Make normalize() always return the command ending in a newline
Except for a few cases that are invalid syntax.

This makes this work better with things like inspect, which expect the file
lines to end in a newline. I could have fixed this in smart_eval, but since
normalize already returns results with newlines at the end in the case of
magics and ?, it is obviously not an issue, so I do it there.