Increment the prompt number when an exception is raised
Otherwise, the source for the prompt is not saved in the linecache, meaning
any functions defined there will not show up properly in tracebacks. It can
also cause issues when pudb exits with an exception, for instance, and the
source wouldn't be in In.

This behavior matches the IPython behavior, where prompt numbers are only not
incremented on empty prompts.