Don't try to include the last line if it is blank for M-;
The logic here is not actually correct (it is looking at the last line of the
buffer rather than the last line of the selection). It isn't actually that
important to get this right, and I'm not even so sure if the behavior we were
going for here *is* right. Actually, if we wanted to match emacs, we should
not add any comment delimiters at all to empty lines at the end of the
selection.