Work around custom keybinding issues in prompt-toolkit 3
We can no longer add things to the Keys class because it is now an immutable
Enum, but prompt-toolkit also accepts single character strings as keys, so we
just pick random characters for the custom keys. As long as we add them to the
ANSI_SEQUENCES dict, they will work.

This also adds a monkeypatch workaround for a regression in recent versions of
prompt-toolkit 3 that make undo work character-by-character.

It also removes the keybinding for Control-/ (undo), since it isn't actually
necessary as long as I don't have any keybinding set for it in iTerm2 (this
might have changed in a recent update to iTerm2 that added better xterm key
compatibility).